feat: generate insomnia-inso docs from source (#8099)
* feat: WIP generate insomnia-inso docs from source

Add a script that generates the reference docs for insomnia-inso from
the source code. Commander exposes commands, options and descriptions
which we can use to generate the markdown files.

This is a WIP to start a convo and an initial implementation, the script
doesn't work as it is because `cli.ts` doesn't export `program` and
`version`.

* wire up program

* simplify

* normalise line endings in each function

* Add page.release to relative links

* Use major.minor when generating the docs

* Add frontmatter to generated docs

* Use /inso-cli/reference/<path>/<version>/ as the url segment in generated docs

* Use `title` and not h1s for titles

* Fix linting issues

---------

Co-authored-by: jackkav <jackkav@gmail.com>