[INS-1234] Include token as backup credential when initially fetching repo contents (#9092)
* [INS-1234] Include token as backup credential when initially fetching repo contents

The 'custom' Git tab was failing because it would send the credentials using the token field while the submit method expected it in the password field. Since the method is reused by multiple code paths the most resilient patch is simply to add the alternative possibility, the token, as a fallback.

* fix: update Git credentials handling to use password instead of token and fix readonly inputs

---------

Co-authored-by: James Gatz <jamesgatzos@gmail.com>